Error on unmatched profile name. Handle profile name env var
Also includes a pretty decent sized refactor that I had to do in order
to be able to add these features (for my brain to work with the
conditionals). There's still some repeated code that could be factored
out, but I didn't want to get sucked into too deep of a refactor before
getting feedback.